댓글 수정

+++ src/main/java/org/likelion/likelion_12th_team05/comment/appication/CommentService.java
@@ -2,50 +2,85 @@ package org.likelion.likelion_12th_team05.comment.appication;
 
 import org.likelion.likelion_12th_team05.comment.api.request.CommentSaveReqDto;
 import org.likelion.likelion_12th_team05.comment.api.request.CommentUpdateReqDto;
+import org.likelion.likelion_12th_team05.comment.api.response.CommentInfoResDto;
 import org.likelion.likelion_12th_team05.comment.domain.Comment;
 import org.likelion.likelion_12th_team05.comment.domain.CommentRepository;
+import org.likelion.likelion_12th_team05.common.EntityFinder;
 import org.likelion.likelion_12th_team05.common.error.ErrorCode;
 import org.likelion.likelion_12th_team05.common.exception.CustomException;
-import org.likelion.likelion_12th_team05.curation.domain.Curation;
+import org.likelion.likelion_12th_team05.common.exception.NotFoundException;
 import org.likelion.likelion_12th_team05.curation.domain.repository.CurationRepository;
+import org.likelion.likelion_12th_team05.user.domain.User;
+import org.likelion.likelion_12th_team05.user.domain.repository.UserRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.security.Principal;
+
 @Service
-@Transactional(readOnly = true)
 public class CommentService {
     private final CommentRepository commentRepository;
-    private final CurationRepository curationRepository;
+    private final UserRepository userRepository;
 
-    public CommentService(CommentRepository commentRepository, CurationRepository curationRepository) {
+    public CommentService(CommentRepository commentRepository, CurationRepository curationRepository, UserRepository userRepository) {
         this.commentRepository = commentRepository;
-        this.curationRepository = curationRepository;
+        this.userRepository = userRepository;
     }
 
-    public void commentSave(CommentSaveReqDto commentSaveReqDto) {
-        Curation curation = curationRepository.findById(commentSaveReqDto.id())
-                .orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND_EXCEPTION, "해당하는 사용자가 없습니다."));
+    @Transactional
+    public CommentInfoResDto commentSave(CommentSaveReqDto commentSaveReqDto, Principal principal) {
+        String email = principal.getName();
+        User user = getUserByEmail(email);
 
-        Comment comment = Comment.builder()
-                .id(commentSaveReqDto.id())
-                .name(commentSaveReqDto.name())
-                .comment(commentSaveReqDto.comment())
-                .curation(curation).build();
-
+        Comment comment = commentSaveReqDto.toEntity(user);
+        user.setCommentCount(user.getCommentCount() + 1);
         commentRepository.save(comment);
+
+        return CommentInfoResDto.from(comment);
     }
 
-    public void commentUpdate(Long curationId, CommentUpdateReqDto commentUpdateReqDto) {
-        Comment comment = commentRepository.findById(curationId)
-                .orElseThrow(() -> new CustomException(ErrorCode.CURATIONS_NOT_FOUND_EXCEPTION, "해당하는 큐레이션이 없습니다."));
+    @Transactional
+    public CommentInfoResDto commentUpdate(Long curationId, CommentUpdateReqDto commentUpdateReqDto, Principal principal) {
+        String email = principal.getName();
+        User user = getUserByEmail(email);
 
-        comment.updateCommentInfo(commentUpdateReqDto);
+        Comment comment = getCommentById(curationId);
+
+        String commentUserEmail = comment.getUser().getEmail();
+        if (!email.equals(commentUserEmail)) {
+            throw new CustomException(ErrorCode.NO_AUTHORIZATION_EXCEPTION,
+                    ErrorCode.NO_AUTHORIZATION_EXCEPTION.getMessage());
+        }
+
+        comment.updateCommentInfo(commentUpdateReqDto, user);
+        commentRepository.save(comment);
+        return CommentInfoResDto.from(comment);
     }
 
-    public void commentDelete(Long commentId) {
-        Comment comment = commentRepository.findById(commentId)
-                .orElseThrow(() -> new CustomException(ErrorCode.CURATIONS_NOT_FOUND_EXCEPTION, "해당하는 큐레이션이 없습니다."));
+    @Transactional
+    public void commentDelete(Long commentId, Principal principal) {
+        Comment comment = getCommentById(commentId);
 
+        String email = principal.getName();
+        User user = getUserByEmail(email);
+        String commentUserEmail = comment.getUser().getEmail();
+
+        if (!email.equals(commentUserEmail)) {
+            throw new NotFoundException(ErrorCode.USER_NOT_FOUND_EXCEPTION,
+                    ErrorCode.USER_NOT_FOUND_EXCEPTION.getMessage());
+        }
+
+        user.setCommentCount(user.getCommentCount() - 1);
         commentRepository.delete(comment);
+    }
+
+    private Comment getCommentById(Long commentId) {
+        return EntityFinder.findByIdOrThrow
+                (commentRepository.findById(commentId), ErrorCode.USER_NOT_FOUND_EXCEPTION);
+    }
+
+    private User getUserByEmail(String email) {
+        return EntityFinder.findByEmailOrThrow
+                (userRepository.findByEmail(email), ErrorCode.USER_NOT_FOUND_EXCEPTION);
     }
 }
